tanker_get_device_list: Make the device count 32 bits, not 64

+++ modules/sdk-c/include/ctanker/ctanker.h
@@ -67,7 +67,7 @@ typedef struct tanker_device_list tanker_device_list_t;
  */
 struct tanker_device_list
 {
-  uint64_t count;
+  uint32_t count;
   tanker_device_list_elem_t* devices;
 };
